refactor(main): encapsulate lotto game logic in LottoGame class
Moved the main lotto game logic from Application.kt to a new LottoGame class, improving code organization and readability. The main function now initializes and runs the LottoGame instance.
